+++ support/kernel/kernel_windows.go
@@ -1,0 +1,18 @@
+// Package kernel provides methods for collecting information from OS kernel
+package kernel
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                         Copyright (c) 2024 ESSENTIAL KAOS                          //
+//      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+import "github.com/essentialkaos/ek/v13/support"
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// Collect collects info from OS kernel
+func Collect(params ...string) []support.KernelParam {
+	return nil
+}